+++ lib/part10_builder_statefulwidget.dart
@@ -17,58 +17,62 @@ class _BuilderState extends State<_BuilderStatefulWidget> {
   void initState() {
     super.initState();
     // Record this State object
-    appState = RootState._rootStateX!;
+    appState = RootStateMixin._rootStateX!;
     appState._builderState = this;
   }
 
   late AppStateX appState;
 
   @override
-  Widget build(BuildContext context) {
-    //
-    Widget? widget;
-    FlutterErrorDetails? errorDetails;
+  Widget build(BuildContext context) => appState.builder(context);
+  // Error handler is introduced in class AppStateX (FlutterError.onError = _errorHandler;)
 
-    try {
-      widget = appState.builder(context);
-    } catch (e) {
-      errorDetails = FlutterErrorDetails(
-        exception: e,
-        stack: e is Error ? e.stackTrace : null,
-        library: 'part10_builder_statefulwidget.dart',
-        context: ErrorDescription('Error with appState.builder()'),
-      );
-    }
-
-    // A error
-    if (errorDetails != null) {
-      //
-      try {
-        appState.onError(errorDetails);
-      } catch (e) {
-        // ignore error
-      }
-
-      try {
-        widget = ErrorWidget.builder(errorDetails);
-      } catch (e) {
-        // Throw in DebugMode.
-        if (kDebugMode) {
-          rethrow;
-        } else {
-          // Record error in log
-          _logPackageError(
-            e,
-            library: 'part10_builder_statefulwidget.dart',
-            description: 'Error with ErrorWidget.builder(errorDetails)',
-          );
-        }
-      }
-    }
-
-    // Must provide something. Blank then
-    widget ??= const SizedBox.shrink();
-
-    return widget;
-  }
+  // @override
+  // Widget build(BuildContext context) {
+  //   //
+  //   Widget? widget;
+  //   FlutterErrorDetails? errorDetails;
+  //
+  //   try {
+  //     widget = appState.builder(context);
+  //   } catch (e) {
+  //     errorDetails = FlutterErrorDetails(
+  //       exception: e,
+  //       stack: e is Error ? e.stackTrace : null,
+  //       library: 'part10_builder_statefulwidget.dart',
+  //       context: ErrorDescription('Error with appState.builder()'),
+  //     );
+  //   }
+  //
+  //   // A error
+  //   if (errorDetails != null) {
+  //     //
+  //     try {
+  //       appState.onError(errorDetails);
+  //     } catch (e) {
+  //       // ignore error
+  //     }
+  //
+  //     try {
+  //       widget = ErrorWidget.builder(errorDetails);
+  //     } catch (e) {
+  //       // Throw in DebugMode.
+  //       if (kDebugMode) {
+  //         rethrow;
+  //       } else {
+  //         // Record error in log
+  //         _logPackageError(
+  //           e,
+  //           library: 'part10_builder_statefulwidget.dart',
+  //           description: 'Error with ErrorWidget.builder(errorDetails)',
+  //         );
+  //       }
+  //     }
+  //   }
+  //
+  //   // Must provide something. Blank then
+  //   widget ??= const SizedBox.shrink();
+  //
+  //   return widget;
+  // }
 }