+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+## 5.10.0
+February 28, 2025
+- In StateX class:  catch (e) { to catch (e, stack) {
+- In InheritedWidgetStateMixin: Replace stateSet() with setBuilder(WidgetBuilder? builder)
+- New getter in AppStateX class: inErrorRoutine
+- Rewrite of class _BuilderState
+- Renamed mixin RootState to RootStateMixin
+
 ## 5.9.0
 December 14 ,2024
 - in part13_statex_controller.dart, call all listeners in both